+++ database/migrations/2020_11_23_105424_create_items_table.php
@@ -16,7 +16,7 @@ class CreateItemsTable extends Migration
         Schema::create('items', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->unsignedInteger('category_id')->index();
+            $table->unsignedBigInteger('category_id')->index();
             $table->integer('initialQty');
             $table->foreign('category_id')->references('id')->on('categories');
             $table->timestamps();
