+++ database/migrations/2020_11_24_060126_create_orders_table.php
@@ -16,7 +16,7 @@ class CreateOrdersTable extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
             $table->unsignedInteger('receiving_id')->index();
-            $table->unsignedInteger('item_id')->index();
+            $table->unsignedBigInteger('item_id')->index();
 
             $table->foreign('receiving_id')->references('id')->on('receivings');
             $table->foreign('item_id')->references('id')->on('items');
